fix(react) - Have individual props for each card
