+++ packages/react/src/components/Cards/DetailCard/DetailCard.args.ts
@@ -1,0 +1,17 @@
+import { DetailCardProps } from ".";
+
+export const DetailCardArgs: DetailCardProps = {
+  title: "Technical meeting on digitalization in the construction sector",
+  eyebrow: "Meeting",
+  intro:
+    "The purpose of the meeting will be to discuss opportunities and challenges for the future of work in the construction industry as a vehicle to ensure a human-centred economic recovery.",
+  date: {
+    human: "18 March 2023",
+    unix: "1670389200",
+  },
+  image: "/medium.jpg",
+  link: "https:/www.ilo.org",
+  eventdetails: "8:30 - 12:00 CET | Geneva",
+};
+
+export default DetailCardArgs;
