+++ packages/react/src/components/Cards/DetailCard/DetailCard.props.ts
@@ -1,0 +1,43 @@
+import { CardSize, EventDate } from "../../../types";
+
+export type DetailCardProps = {
+  /**
+   * A line of text that appears as a small heading above the title of the card
+   */
+  eyebrow?: string;
+
+  /**
+   * The title of the card
+   */
+  title: string;
+
+  /**
+   * Introductory text in the card
+   */
+  intro?: string;
+
+  /**
+   * Specify the event Date, in both human and Unix format.
+   */
+  date?: EventDate;
+
+  /**
+   * Information about an event
+   */
+  eventdetails?: string;
+
+  /**
+   * A URL to link to
+   */
+  link?: string;
+
+  /**
+   * The image to show in the card
+   */
+  image?: string;
+
+  /**
+   * How big should the card be
+   */
+  size?: Omit<CardSize, "standard">;
+};
