+++ packages/react/src/components/Cards/FactlistCard/FactListCard.args.ts
@@ -1,0 +1,14 @@
+import { FactlistCardProps } from ".";
+
+export const FactlistCardArgs: FactlistCardProps = {
+  title:
+    "Economic slowdown likely to force workers to accept lower quality jobs",
+  theme: "light",
+  list: [
+    "Global employment growth will be only 1.0 per cent in 2023, less than half the level in 2022.",
+    "The labour market deterioration is mainly due to emerging geopolitical tensions and the Ukraine conflict.",
+    "The current slowdown means that many workers will have to accept lower quality jobs, often at very low pay.",
+  ],
+};
+
+export default FactlistCardArgs;
