+++ packages/react/src/components/Cards/MultilinkCard/MultilinkCard.args.ts
@@ -1,0 +1,33 @@
+import { MultilinkCardProps } from ".";
+
+export const MultilinkCardArgs: MultilinkCardProps = {
+  title:
+    "ILO welcomes G7 call to make a just transition to a green economy happen",
+  eyebrow: "High-level meeting",
+  intro:
+    "At the end of their meeting the G7 Labour Ministers highlighted the urgent need to greater focus on rights and occupational safety and health.",
+  image: "/hero.jpg",
+  alignment: "left",
+  link: "https:/www.ilo.org",
+  linklist: {
+    headline: "",
+    linkgroup: [
+      {
+        links: [
+          {
+            label: "Read the press release",
+            url: "https://www.ilo.org",
+          },
+          {
+            label: "See the statement",
+            url: "https://www.ilo.org",
+          },
+          {
+            label: "Remarks to G7 Openening Session",
+            url: "https://www.ilo.org",
+          },
+        ],
+      },
+    ],
+  },
+};
