+++ packages/react/src/components/Cards/MultilinkCard/MultilinkCard.props.ts
@@ -1,0 +1,30 @@
+import { CardAlignment, CardSize } from "../../../types";
+import { LinkListProps } from "../../LinkList/LinkList.props";
+
+export type MultilinkCardProps = {
+  isvideo?: boolean;
+
+  /**
+   * A line of text that appears as a small heading above the title of the card
+   */
+  eyebrow?: string;
+
+  /**
+   * The title of the card
+   */
+  title: string;
+
+  /**
+   * How big should the card be
+   */
+  size?: CardSize;
+  alignment?: CardAlignment;
+  intro?: string;
+  link?: string;
+  linklist?: LinkListProps;
+
+  /**
+   * The image to show in the card
+   */
+  image?: string;
+};
