+++ packages/react/src/components/Cards/PromoCard/PromoCard.args.ts
@@ -1,0 +1,19 @@
+import { PromoCardProps } from ".";
+
+export const PromoCardArgs: PromoCardProps = {
+  size: "standard",
+  title:
+    "ILO welcomes G7 call to make a just transition to a green economy happen",
+  eyebrow: "High-level meeting",
+  theme: "dark",
+  cornercut: true,
+  intro:
+    "At the end of their meeting the G7 Labour Ministers highlighted the urgent need to greater focus on rights and occupational safety and health.",
+  link: "https:/www.ilo.org",
+  cta: {
+    label: "Read the press release",
+    url: "https://www.ilo.org",
+  },
+};
+
+export default PromoCardArgs;
