+++ packages/react/src/components/Cards/PromoCard/PromoCard.props.ts
@@ -1,0 +1,36 @@
+import { ThemeTypes, CardCornerType, CardSize } from "../../../types";
+import { LinkProps } from "../../LinkList/LinkList.props";
+
+export type PromoCardProps = {
+  /**
+   * A line of text that appears as a small heading above the title of the card
+   */
+  eyebrow?: string;
+
+  /**
+   * The title of the card
+   */
+  title: string;
+
+  /**
+   * Will render the card to appear on light or dark backgrounds
+   */
+  theme?: ThemeTypes;
+
+  /**
+   * How big should the card be
+   */
+  size?: CardSize;
+
+  /**
+   * Apply an optional corner cut to the top of the card
+   */
+  cornercut?: CardCornerType;
+
+  /**
+   * Introductory text in the card
+   */
+  intro?: string;
+  link?: string;
+  cta?: LinkProps;
+};
