+++ packages/react/src/components/Cards/StatCard/StatCard.args.ts
@@ -1,0 +1,14 @@
+import { StatCardProps } from ".";
+
+export const StatCardArgs: StatCardProps = {
+  title: "Global employment growth down by half in 2023",
+  intro:
+    "The current global economic slowdown is likely to force more workers to accept lower quality, poorly paid jobs which lack job security and social protection, so accentuating inequalities exacerbated by the COVID-19 crisis.",
+  color: "turquoise",
+  source: {
+    label: "World Employment and Social Outlook: Trends 2023",
+    url: "https://www.ilo.org",
+  },
+};
+
+export default StatCardArgs;
