+++ packages/react/src/components/Cards/TextCard/TextCard.args.ts
@@ -1,0 +1,21 @@
+import { TextCardProps } from ".";
+
+export const TextCardArgs: TextCardProps = {
+  title: "Why we need greater social justice",
+  eyebrow: "Podcast",
+  date: {
+    human: "20 September 2022",
+    unix: "1670389200",
+  },
+  theme: "light",
+  link: "https:/www.ilo.org",
+  profile: {
+    avatar: "/ilo-dg.jpg",
+    description:
+      "Gilbert Houngbo is the Director-General of the International Labour Organization",
+    name: "Gilbert Houngbo",
+    role: "ILO Director-General",
+  },
+};
+
+export default TextCardArgs;
