--- packages/react/src/components/Cards/index.ts
@@ -1,21 +1,0 @@
-import FeatureCard from "./FeatureCard";
-import DataCard from "./DataCard";
-import DetailCard from "./DetailCard";
-import FactlistCard from "./FactlistCard";
-import MultilinkCard from "./MultilinkCard";
-import PromoCard from "./PromoCard";
-import StatCard from "./StatCard";
-import TextCard from "./TextCard";
-
-export {
-  FeatureCard,
-  DataCard,
-  DetailCard,
-  FactlistCard,
-  MultilinkCard,
-  PromoCard,
-  StatCard,
-  TextCard,
-};
-
-export { default as CardGroup } from "./CardGroup";
